+++ src/components/TheFooter.vue
@@ -66,7 +66,7 @@
 </template>
 
 <script lang="ts" setup>
-import { canAdmin } from '../utils/currentUserCan';
+import { canAdmin } from '../utils/user';
 import { useI18n } from 'vue-i18n';
 import { generalContent } from '../store';
