+++ src/components/pages/profile/contribution/CancelContribution.vue
@@ -26,7 +26,7 @@
 <script lang="ts" setup>
 import { ContributionType, MembershipStatus } from '@beabee/beabee-common';
 import AppButton from '../../../button/AppButton.vue';
-import { formatLocale } from '../../../../utils/dates/locale-date-formats';
+import { formatLocale } from '../../../../utils/dates';
 import { computed } from 'vue';
 import { useI18n } from 'vue-i18n';
 import { ContributionInfo } from '../../../../utils/api/api.interface';
