+++ src/components/pages/profile/contribution/ProrateContribution.vue
@@ -59,7 +59,7 @@
 import { differenceInMonths } from 'date-fns';
 import { computed, ref, watch } from 'vue';
 import { useI18n } from 'vue-i18n';
-import { formatLocale } from '../../../../utils/dates/locale-date-formats';
+import { formatLocale } from '../../../../utils/dates';
 
 const { t, n } = useI18n();
