+++ src/components/pages/profile/contribution/UpdateContribution.vue
@@ -109,7 +109,7 @@ import AppModal from '../../../AppModal.vue';
 import StripePayment from '../../../StripePayment.vue';
 import { currentUser } from '../../../../store/currentUser';
 import AppAlert from '../../../AppAlert.vue';
-import { formatLocale } from '../../../../utils/dates/locale-date-formats';
+import { formatLocale } from '../../../../utils/dates';
 import AppHeading from '../../../AppHeading.vue';
 import { isRequestError } from '../../../../utils/api';
 import { ContributionInfo } from '../../../../utils/api/api.interface';
