+++ src/components/search/operators/ArgInput.vue
@@ -31,7 +31,7 @@
 <script lang="ts" setup>
 import { RuleValue } from '@beabee/beabee-common';
 import { computed } from 'vue';
-import { formatLocale } from '../../../utils/dates/locale-date-formats';
+import { formatLocale } from '../../../utils/dates';
 import AppInput from '../../forms/AppInput.vue';
 import AppRadioGroup from '../../forms/AppRadioGroup.vue';
 import AppSelect from '../../forms/AppSelect.vue';
