+++ src/pages/admin/callouts/view/[id]/responses/[rid].vue
@@ -126,7 +126,7 @@ import { useI18n } from 'vue-i18n';
 import AppHeading from '../../../../../../components/AppHeading.vue';
 import AppInfoList from '../../../../../../components/AppInfoList.vue';
 import AppInfoListItem from '../../../../../../components/AppInfoListItem.vue';
-import { formatLocale } from '../../../../../../utils/dates/locale-date-formats';
+import { formatLocale } from '../../../../../../utils/dates';
 import AppButton from '../../../../../../components/button/AppButton.vue';
 import AppButtonGroup from '../../../../../../components/button/AppButtonGroup.vue';
 import { addBreadcrumb } from '../../../../../../store/breadcrumb';
