+++ src/pages/auth/login.vue
@@ -68,7 +68,7 @@ import { useRoute } from 'vue-router';
 import { useI18n } from 'vue-i18n';
 import useVuelidate from '@vuelidate/core';
 import { reactive, ref } from 'vue';
-import isInternalUrl from '../../utils/is-internal-url';
+import { isInternalUrl } from '../../utils';
 import { updateCurrentUser } from '../../store';
 import { LoginData } from '../../utils/api/api.interface';
 import { login } from '../../utils/api/auth';
