--- src/utils/is-internal-url.ts
@@ -1,4 +1,0 @@
-function isInternalUrl(url: string | undefined): url is string {
-  return !!url && /^\/([^/]|$)/.test(url);
-}
-export default isInternalUrl;
